Rename clear logs, update icon #467
