+++ cogboard-webapp/src/components/widgets/types/LogViewerWidget/Toolbar/DateRangePicker/index.js
@@ -1,12 +1,16 @@
 import React from 'react';
+import moment from 'moment-timezone';
 import MomentUtils from '@date-io/moment';
 import { getDateSpan, saveDateSpan } from './helpers';
 
-import ToolbarGroup from '../ToolbarGroup';
 import { MuiPickersUtilsProvider } from '@material-ui/pickers';
+import UpdateIcon from '@material-ui/icons/Update';
+import ToolbarGroup from '../ToolbarGroup';
+import ToggleIconButton from '../ToggleIconButton';
 import CustomPicker from './CustomPicker';
+import { DatePickerWrapper } from './styled';
 
-const DateRangePicker = ({ widgetLocalStorage }) => {
+const DateRangePicker = ({ widgetLocalStorage, lastLog }) => {
   const { begin, end } = getDateSpan(widgetLocalStorage);
 
   const handleBeginChange = date =>
@@ -14,14 +18,30 @@ const DateRangePicker = ({ widgetLocalStorage }) => {
   const handleEndChange = date =>
     saveDateSpan(widgetLocalStorage, { begin, end: date });
 
+  const handleClearLogs = () => {
+    const date = lastLog?.date;
+    if (date) {
+      const beginDate = moment(date).add(1, 'seconds');
+      saveDateSpan(widgetLocalStorage, { begin: beginDate, end: null });
+    }
+  };
+
   return (
     <ToolbarGroup title="Date span">
       <MuiPickersUtilsProvider utils={MomentUtils}>
-        <CustomPicker
-          label="Begin"
-          value={begin}
-          onChange={handleBeginChange}
-        />
+        <DatePickerWrapper>
+          <CustomPicker
+            label="Begin"
+            value={begin}
+            onChange={handleBeginChange}
+          />
+          <ToggleIconButton
+            tooltip="Show logs from now"
+            data-cy="show-logs-from-now-button"
+            onClick={handleClearLogs}
+            Icon={UpdateIcon}
+          />
+        </DatePickerWrapper>
         <CustomPicker label="End" value={end} onChange={handleEndChange} />
       </MuiPickersUtilsProvider>
     </ToolbarGroup>
