+++ cogboard-webapp/src/components/widgets/types/LogViewerWidget/Toolbar/index.js
@@ -1,6 +1,4 @@
 import React from 'react';
-import moment from 'moment-timezone';
-import { saveDateSpan } from './DateRangePicker/helpers';
 
 import ToggleIconButton from './ToggleIconButton';
 import { Wrapper } from './styled';
@@ -8,7 +6,6 @@ import SearchInput from './SearchInput';
 import ToolbarGroup from './ToolbarGroup';
 import DateRangePicker from './DateRangePicker';
 import ArrowDownwardIcon from '@material-ui/icons/ArrowDownward';
-import DeleteIcon from '@material-ui/icons/Delete';
 import FilterPicker from './FilterPicker';
 
 const Toolbar = ({
@@ -20,14 +17,6 @@ const Toolbar = ({
   handleFollowChange,
   lastLog
 }) => {
-  const handleClearLogs = () => {
-    const date = lastLog?.date;
-    if (date) {
-      const beginDate = moment(date).add(1, 'seconds');
-      saveDateSpan(widgetLocalStorage, { begin: beginDate, end: null });
-    }
-  };
-
   return (
     <Wrapper>
       <ToolbarGroup>
@@ -44,7 +33,10 @@ const Toolbar = ({
         quarantine={quarantine}
       />
 
-      <DateRangePicker widgetLocalStorage={widgetLocalStorage} />
+      <DateRangePicker
+        widgetLocalStorage={widgetLocalStorage}
+        lastLog={lastLog}
+      />
 
       <ToolbarGroup>
         <ToggleIconButton
@@ -52,12 +44,6 @@ const Toolbar = ({
           onClick={() => handleFollowChange(!shouldFollowLogs)}
           enabled={shouldFollowLogs}
           Icon={ArrowDownwardIcon}
-        />
-        <ToggleIconButton
-          tooltip="Clear logs"
-          data-cy="clear-logs-button"
-          onClick={handleClearLogs}
-          Icon={DeleteIcon}
         />
       </ToolbarGroup>
     </Wrapper>
